reduce size models, increase perfomance on devices
