--- public/green_tea_vending_machine/license.txt
@@ -1,11 +1,0 @@
-Model Information:
-* title:	Green Tea Vending Machine
-* source:	https://sketchfab.com/3d-models/green-tea-vending-machine-7fb6aca9083d40b3acf03dd6de71f90f
-* author:	JosevanD (https://sketchfab.com/JosevanD2)
-
-Model License:
-* license type:	CC-BY-NC-4.0 (http://creativecommons.org/licenses/by-nc/4.0/)
-* requirements:	Author must be credited. No commercial use.
-
-If you use this 3D model in your project be sure to copy paste this credit wherever you share it:
-This work is based on "Green Tea Vending Machine" (https://sketchfab.com/3d-models/green-tea-vending-machine-7fb6aca9083d40b3acf03dd6de71f90f) by JosevanD (https://sketchfab.com/JosevanD2) licensed under CC-BY-NC-4.0 (http://creativecommons.org/licenses/by-nc/4.0/)
