+++ public/planet_mobile/scene.gltf
@@ -1,0 +1,1 @@
+{"asset":{"generator":"glTF-Transform v3.4.0","version":"2.0","extras":{"author":"cmzw (https://sketchfab.com/cmzw)","license":"CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)","source":"https://sketchfab.com/3d-models/stylized-planet-789725db86f547fc9163b00f302c3e70","title":"Stylized planet"}},"accessors":[{"type":"SCALAR","componentType":5125,"count":103014},{"type":"VEC3","componentType":5126,"count":26576},{"type":"VEC3","componentType":5126,"count":26576,"max":[0.8758764863014221,0.8545469045639038,0.8728971481323242],"min":[-0.8763356804847717,-0.8634992837905884,-0.8758782148361206]},{"type":"VEC2","componentType":5126,"count":26576},{"type":"SCALAR","componentType":5125,"count":39042},{"type":"VEC3","componentType":5126,"count":6587},{"type":"VEC3","componentType":5126,"count":6587,"max":[0.6551179885864258,0.6613333821296692,0.6618664860725403],"min":[-0.6606217622756958,-0.6490849852561951,-0.6526646018028259]},{"type":"VEC2","componentType":5126,"count":6587},{"type":"SCALAR","componentType":5126,"count":4,"max":[15],"min":[0],"bufferView":0,"byteOffset":0},{"type":"VEC4","componentType":5126,"count":4,"bufferView":0,"byteOffset":16},{"type":"SCALAR","componentType":5126,"count":2,"max":[15],"min":[0],"bufferView":0,"byteOffset":80},{"type":"VEC3","componentType":5126,"count":2,"bufferView":0,"byteOffset":88}],"bufferViews":[{"buffer":0,"byteOffset":0,"byteLength":112},{"buffer":0,"byteOffset":112,"byteLength":150077},{"buffer":0,"byteOffset":150189,"byteLength":48854}],"samplers":[{"magFilter":9729,"minFilter":9987,"wrapS":10497,"wrapT":10497}],"textures":[{"source":0,"sampler":0},{"source":1,"sampler":0}],"images":[{"mimeType":"image/png","uri":"textures/Clouds_baseColor.png"},{"mimeType":"image/png","uri":"textures/Planet_baseColor.png"}],"buffers":[{"uri":"scene.bin","byteLength":199043}],"materials":[{"name":"Clouds","doubleSided":true,"pbrMetallicRoughness":{"metallicFactor":0,"baseColorTexture":{"index":0}},"emissiveTexture":{"index":0},"extensions":{"KHR_materials_unlit":{}}},{"name":"Planet","pbrMetallicRoughness":{"metallicFactor":0,"baseColorTexture":{"index":1}},"emissiveFactor":[0.23391156271636818,0.23391156271636818,0.23391156271636818],"emissiveTexture":{"index":1},"extensions":{"KHR_materials_unlit":{}}}],"meshes":[{"name":"Object_0","primitives":[{"attributes":{"NORMAL":1,"POSITION":2,"TEXCOORD_0":3},"mode":4,"material":0,"indices":0,"extensions":{"KHR_draco_mesh_compression":{"bufferView":1,"attributes":{"NORMAL":0,"POSITION":1,"TEXCOORD_0":2}}}}]},{"name":"Object_1","primitives":[{"attributes":{"NORMAL":5,"POSITION":6,"TEXCOORD_0":7},"mode":4,"material":1,"indices":4,"extensions":{"KHR_draco_mesh_compression":{"bufferView":2,"attributes":{"NORMAL":0,"POSITION":1,"TEXCOORD_0":2}}}}]}],"nodes":[{"name":"Sketchfab_model","rotation":[-0.6956252833973351,-0.022898241408018643,0.02219405001814219,0.7176967046321006],"children":[1]},{"name":"root","children":[2]},{"name":"GLTF_SceneRootNode","rotation":[0.7071067811865475,0,0,0.7071067811865476],"children":[3,5]},{"name":"Clouds_1","children":[4]},{"name":"Object_4","mesh":0},{"name":"Planet_2","children":[6]},{"name":"Object_6","mesh":1}],"animations":[{"name":"Animation","samplers":[{"input":8,"output":9,"interpolation":"LINEAR"},{"input":10,"output":11,"interpolation":"LINEAR"},{"input":8,"output":9,"interpolation":"LINEAR"},{"input":10,"output":11,"interpolation":"LINEAR"}],"channels":[{"sampler":0,"target":{"node":3,"path":"rotation"}},{"sampler":1,"target":{"node":3,"path":"scale"}},{"sampler":2,"target":{"node":5,"path":"rotation"}},{"sampler":3,"target":{"node":5,"path":"scale"}}]}],"scenes":[{"name":"Sketchfab_Scene","nodes":[0]}],"scene":0,"extensionsUsed":["KHR_materials_unlit","KHR_draco_mesh_compression"],"extensionsRequired":["KHR_draco_mesh_compression"]}
