+++ public/planet_pc/scene.gltf
