--- src/components/canvas/Ball.tsx
@@ -1,77 +1,0 @@
-import React, { Suspense } from "react";
-import { Canvas } from "@react-three/fiber";
-import {
-  Decal,
-  Float,
-  OrbitControls,
-  Preload,
-  useTexture,
-} from "@react-three/drei";
-
-import CanvasLoader from "@components/Loader";
-
-export const Demo = (props) => {
-  const [decal] = useTexture([props.imgUrl]);
-  return (
-      <mesh castShadow={true} receiveShadow={true} scale={2.75}>
-        <icosahedronGeometry args={[1, 3]} />
-        <meshStandardMaterial
-            color='#fff8eb'
-            polygonOffset={true}
-            polygonOffsetFactor={-5}
-            flatShading={true}
-        />
-        <Decal
-            position={[0, 0, 1]}
-            rotation={[2 * Math.PI, 0, 6.25]}
-            scale={1}
-            map={decal}
-        />
-      </mesh>
-  )
-}
-
-export const Ball = (props) => {
-  const [decal] = useTexture([props.imgUrl]);
-
-  return (
-    <Float speed={1.75} rotationIntensity={1} floatIntensity={2}>
-      <ambientLight intensity={0.25} />
-      <directionalLight position={[0, 0, 0.05]} />
-      <mesh castShadow={true} receiveShadow={true} scale={2.75}>
-        <icosahedronGeometry args={[1, 3]} />
-        <meshStandardMaterial
-          color='#fff8eb'
-          polygonOffset={true}
-          polygonOffsetFactor={-5}
-          flatShading={true}
-        />
-        <Decal
-          position={[0, 0, 1]}
-          rotation={[2 * Math.PI, 0, 6.25]}
-          scale={1}
-          map={decal}
-        />
-      </mesh>
-    </Float>
-  );
-};
-
-const BallCanvas = ({ icon }) => {
-  return (
-    <Canvas
-      frameloop='demand'
-      dpr={[1, 2]}
-      gl={{ preserveDrawingBuffer: true }}
-    >
-      <Suspense fallback={<CanvasLoader />}>
-        <OrbitControls enableZoom={false} />
-        <Ball imgUrl={icon} />
-      </Suspense>
-
-      <Preload all />
-    </Canvas>
-  );
-};
-
-export default BallCanvas;
