+++ src/components/canvas/Computer.tsx
@@ -4,14 +4,15 @@ import {OrbitControls, Preload, useGLTF} from "@react-three/drei";
 
 import CanvasLoader from "@components/Loader.tsx";
 
-const Computer = ({isMobile}: {isMobile: boolean}) => {
+const Computer = ({isMobile}: { isMobile: boolean }) => {
     const computer = useGLTF(`./pc_${isMobile ? 'mobile' : 'desktop'}/scene.gltf`)
     return (
         <mesh>
             <hemisphereLight intensity={0.15} groundColor={"black"}/>
             <pointLight intensity={1}/>
             {/*<spotLight position={[0, 0, 0]} angle={0.12} penumbra={1} intensity={1} castShadow={true} shadow-mapSize={1024}/>*/}
-            <primitive position={isMobile ? [-1, -2, 0] : [-1,-4,0]} rotation={[0,2,0]} object={computer.scene} scale={isMobile ? 1 : 2}/>
+            <primitive position={isMobile ? [-1, -2, 0] : [-1, -4, 0]} rotation={[0, 2, 0]} object={computer.scene}
+                       scale={isMobile ? 1 : 2}/>
         </mesh>
     );
 };
@@ -50,7 +51,7 @@ const ComputerCanvas = () => {
                 <Computer isMobile={isMobile}/>
             </Suspense>
 
-            <Preload all />
+            <Preload all/>
         </Canvas>
     )
 }
